TelkomAthon2/requirements.txt
# Core
fastapi==0.111.0
uvicorn==0.29.0
streamlit==1.35.0
python-dotenv==1.0.1
pydantic==2.7.1
python-multipart==0.0.9

# Auth
python-jose[cryptography]==3.3.0
bcrypt==4.1.3

# Azure OpenAI
openai==1.30.1

# Document Parsing
PyMuPDF==1.24.3
python-pptx==0.6.23
python-docx==1.1.2
openpyxl==3.1.2

# Database
sqlalchemy==2.0.30
psycopg2-binary>=2.9.6
supabase==2.4.6
pgvector==0.2.5

# Utilities
pandas==2.2.2
numpy==1.26.4
httpx==0.27.0
